xcsp3 outputo
